feat: move get stable asset
