+++ src/console/src/storage/mod.rs
@@ -1,6 +1,4 @@
 pub mod certified_assets;
-mod impls;
 pub mod state;
 pub mod store;
 pub mod strategy_impls;
-pub mod types;
