+++ src/libs/cdn/src/lib.rs
@@ -1,5 +1,9 @@
 mod storage;
 pub mod strategies;
+mod types;
 
 pub use storage::certified_assets::init_certified_assets;
 pub use storage::state::heap;
+pub use storage::state::stable;
+pub use storage::types::state::*;
+pub use types::state::*;
