+++ src/libs/cdn/src/storage/mod.rs
@@ -1,2 +1,3 @@
 pub mod certified_assets;
 pub mod state;
+pub mod types;
