+++ src/libs/cdn/src/storage/types.rs
