+++ src/libs/cdn/src/types.rs
@@ -1,0 +1,3 @@
+pub mod state {
+    pub type ProposalId = u128;
+}
